hospital-pequeno/requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
xhtml2pdf==0.2.13
openpyxl==3.1.2
Pillow==10.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0
